Move Tpu related logic to a seperate folder
