+++ pkg/commonutils/statefulset/statefulset_utils.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package commonutils
+package statefulsetutils
 
 import (
 	"regexp"
